api-server/water-forecast-master/requirements.txt
joblib>=1.0.0
tensorflow==2.18.1
scikit_optimize==0.9.0
pmdarima==1.7.1
kmodes==0.10.2
pandas==1.3.3
tqdm==4.60.0
seaborn==0.11.0
dill==0.3.8
matplotlib==3.4.3
statsmodels==0.11.1
prophet==1.1
holidays==0.24
PyYAML==5.2
scikit_learn==1.1.3
numpy==1.21.4
protobuf==3.20
